% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3rr.R
\name{directionForNeighbour}
\alias{directionForNeighbour}
\title{Direction for neighbour
ı}
\usage{
directionForNeighbour(origin, destination)
}
\description{
Direction for neighbour
ı
}
\examples{

directionForNeighbour("89bf4ac0cd7ffff", "89bf4ac0c8bffff")

}
