% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3rr.R
\name{ll2Cell}
\alias{ll2Cell}
\title{ll2Cel}
\usage{
ll2Cell(lat, lon, res)
}
\description{
LatLon to Cell
}
\examples{

ll2Cell(
  lat = c(-37.820197, -37.818476)
  , lon = c(144.983324, 144.967354)
  , res = 12L
  )

}
